refactor(positions): Use integer coordinates instead of floats
- Changed position type from float to int in models.py (TeamData and PositionUpdate)
- Updated both baseline and TT routes to round() position values when saving
- Cleaned up existing team files to use integer coordinates (removes .0 decimals)

Rationale: Screen pixel coordinates don't need decimal precision. Integers are
cleaner, more readable in markdown files, and avoid floating-point artifacts
from drag/auto-align calculations.

All 286 backend tests and 478 frontend tests passing.

+++ backend/models.py
@@ -35,7 +35,7 @@ class TeamData(BaseModel):
     line_manager: str | None = None  # For current org structure
     product_line: str | None = None  # For Product Lines view (Baseline only)
     business_stream: str | None = None  # For Business Streams view (Baseline only)
-    position: dict[str, float] | None = {"x": 0, "y": 0}
+    position: dict[str, int] | None = {"x": 0, "y": 0}
     metadata: dict[str, Any] | None = {}  # Can include align_hint_x ('left'/'center'/'right') and align_hint_y ('top'/'bottom') - optional positioning hints for auto-align
     # Team API fields (optional, for TT Design teams)
     team_api: TeamAPI | None = None
@@ -65,8 +65,8 @@ class TeamData(BaseModel):
 
 class PositionUpdate(BaseModel):
     """Model for updating team position on canvas"""
-    x: float
-    y: float
+    x: int
+    y: int
 
 
 # Snapshot models
